+++ src/messages/link.rs
@@ -5,8 +5,8 @@ use super::{MessageParser, MessageData};
 pub struct LinkMessage {
     source: String,
     target: String,
-    time: usize,
-    id: usize,
+    time: i64,
+    id: i64,
     title: String,
     description: String,
     url: String,
@@ -21,11 +21,11 @@ impl MessageData for LinkMessage {
         &self.target
     }
 
-    fn time(&self) -> usize {
+    fn time(&self) -> i64 {
         self.time
     }
 
-    fn id(&self) -> usize {
+    fn id(&self) -> i64 {
         self.id
     }
 }
@@ -46,8 +46,8 @@ impl MessageParser for LinkMessage {
         LinkMessage {
             source: source.string(),
             target: target.string(),
-            id: id.number() as usize,
-            time: time.number() as usize,
+            id: id.number() as i64,
+            time: time.number() as i64,
             title: title.string(),
             description: description.string(),
             url: url.string(),
